api/requirements.txt
tensorflow
pillow
scikit-image
scikit-learn
joblib
fastapi
fastapi-cors
python-multipart